rename modules with models to make all (except base.py) plural
