+++ api/models/log_events.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 from api.models.base import InternalModelWithName
-from api.models.group import Group
+from api.models.groups import Group
 from api.models.people import Coordinator, Student, Teacher
 
 # We could have created one table listing all possible names of log events, but that might look
